backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
qrcode[pil]==7.4.2
python-dotenv==1.0.0
aiofiles==23.2.1
apscheduler==3.10.4
